x86/idt_entry: Use %rbp in atomic_entry, not %rbx.
On atomic entries, we save all scratch registers, and additionally,
%rbp. Due to programmer error, %rbx (which is not saved) was used in
atomic_entry and atomic_exit instead of %rbp. This corrupts the value
of the %rbx register in code struck by NMI, leading to incorrect
behavior.

This commit fixes the bug in atomic_entry and atomic_exit, so that
interrupted code does not have its state corrupted.