KEvent: always wait for lock after schedule.
Also don't unlock around sched_wake.  If we unlock around sched_wake or
exit early without waiting for the lock in KEvent::wait, the very common
pattern of storing a KEvent on the stack, waiting on it, and setting the
event from a callback breaks because the KEvent is possibly destroyed
before KEvent::set is done using it.