boot: move map_start, map_end assignment out of equality check in init_memory
Previously, the code in question looked like this:
	if (map_start != map_end) {
		map_hhdm_range_pa(map_start, map_end);
		map_start = start;
		map_end = end;
	}
This commit modifies the code such that it now looks like this:
	if (map_start != map_end)
		map_hhdm_range_pa(map_start, map_end);

	map_start = start;
	map_end = end;
map_start and map_end should unconditionally be assigned to.  This bug
would've manifested if the first MB2_MEMMAP_USABLE memory region in the
bootloader-supplied memory map started somewhere other than zero, with
the effect of preventing any HHDM memory mapping.