Ke: add KeSpinlock, KeDPCSpinlock, KeIRQSpinlock types
KeSpinlock is the most lightweight of these spinlock types. It only
disables preemption before taking the lock. KeSpinlock should be
preferred over the other spinlocks whenever possible, as it doesn't
defer DPCs or IRQs.

KeDPCSpinlock is a good middle-ground between KeSpinlock and
KeIRQSpinlock, masking DPCs when it is held.

KeIRQSpinlock is the heaviest spinlock that we provide. It should be
used when synchronization is needed for interrupt contexts.