asm: add rdtsc.h header with rdtsc() and rdtsc_strong() functions.
The function

	unsigned long long rdtsc();

reads the value of the Time-Stamp Counter on the current processor.  The
other function added by this patch,

	unsigned long long rdtsc_strong();

similarily reads the Time-Stamp Counter on the current processor, but
first issues the LFENCE instruction.  According to Linux, the TSC
behaves as if it is some memory location which is being continuously
updated with the current time by some external CPU.  TSC reads can thus
be executed out-of-order w.r.t. preceding instructions.  Issuing LFENCE
before reading the TSC prevents this reordering.