Ke: add KTIMERs, KeSetTimer and KeUnsetTimer functions.
	struct KTIMER { /* ... */ };

is a datatype that represents a kernel timer object. Each CPU keeps a
queue of its own KTIMERs, and KTIMERs are dispatched in a deferred
procedure call.

The following functions are used for managing KTIMERs:

	void KTIMER::init(KTIMER_CALLBACK);

Initializes a KTIMER with a pointer to a callback function.

	bool KeSetTimer(KTIMER *, nsec_t);

Registers a KTIMER on the current CPU's timer queue. If the KTIMER was
already registered, this function is a no-op and returns false.

	bool KeUnsetTimer(KTIMER *);

Removes a KTIMER from the current CPU's timer queue. If the KTIMER was
not registered on a queue, this function is a no-op and returns false.